pyproject.toml
[project]
name = "kismet-fortune"
version = "0.1.0"
description = "An evolutionary take on fortune-mod."
authors = [
    {name = "Steven Miklovic",email = "steven@thinkingsage.com"}
]
license = {text = "GPL-3.0"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer>=0.15.1,<0.16.0",
    "prompt-toolkit>=3.0.48,<4.0.0",
    "google-genai>=0.5.0,<0.6.0",
]

[project.scripts]
kismet = "kismet_fortune.cli:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
